requirements.txt
Django==2.1.7
djangorestframework==3.9.2
# psycopg2-binary>=2.8
psycopg2-binary
PyJWT==1.7.1
requests==2.21.0